requirements.txt
fastapi
uvicorn
alembic
asyncpg
sqlalchemy
pydantic
python-jose
pyjwt
pydantic[email]
pydantic-settings
bcrypt
python-multipart
boto3
celery[redis]
redis
eventlet
fastapi-cache2
